move pairing code to create handler
